refactor: remove spaces from directory names, chore: ignore .env file - Closes #8
