--- .env
@@ -1,43 +1,0 @@
-# Server config
-NODE_ENV=development
-PORT=3000
-
-# Database Config
-DB_HOST=localhost
-DB_PORT=5432
-DB_NAME=acocie_stores
-DB_USER=postgres
-DB_PASSWORD=Kundan1234
-
-# JWT Configuration
-JWT_SECRET=DMPOm4DUtYLU7tuKJkFx0myWh/hMduNf8jwJViwQjRk=
-JWT_REFRESH_SECRET=0s7aGgB2AAMhA/lPtXxuAsptrxm1nmE0KCjGhQc5vMc=
-JWT_ACCESS_EXPIRY=15m
-JWT_REFRESH_EXPIRY=7d
-
-# Email Configuration (SMTP)
-EMAIL_HOST=smtp.gmail.com
-EMAIL_PORT=587
-EMAIL_USER=ondeyostephen0@gmail.com
-EMAIL_PASSWORD=ezfj jpbq qufz eivd
-EMAIL_FROM="Acocie Stores <noreply@acocie.com>"
-
-# Redis Configuration
-REDIS_HOST=localhost
-REDIS_PORT=6379
-REDIS_PASSWORD=
-
-# Frontend URL (for CORS and email links)
-FRONTEND_URL=http://localhost:3001
-
-# Rate Limiting
-RATE_LIMIT_WINDOW_MS=900000
-RATE_LIMIT_MAX_REQUESTS=100
-
-# OTP Configuration
-OTP_EXPIRY_MINUTES=10
-
-# File Upload
-MAX_FILE_SIZE=5242880
-ALLOWED_FILE_TYPES=image/jpeg,image/png,image/gif
-
