+++ src/routes/auth-routes/userRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const userController = require('../../controllers/auth controllers/userController');
+const userController = require('../../controllers/auth-controllers/userController');
 const authenticate = require('../../middleware/authMiddleware');
 
 // All routes require authentication
