+++ src/routes/cart-routes/cartRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const cartController = require('../../controllers/Cart controllers/cartController');
+const cartController = require('../../controllers/cart-controllers/cartController');
 const authenticate = require('../../middleware/authMiddleware');
 
 const optionalAuth = (req,res,next) => {
