+++ src/routes/category-routes/categoryRoutes.js
@@ -1,7 +1,7 @@
 
 const express = require('express');
 const router = express.Router();
-const categoryController = require('../../controllers/category controllers/categoryController');
+const categoryController = require('../../controllers/category-controllers/categoryController');
 const authenticate = require('../../middleware/authMiddleware');
 const { isAdmin } = require('../../middleware/roleMiddleware');
