+++ src/routes/checkout-routes/checkoutRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const checkoutController = require('../../controllers/checkout controllers/checkoutController');
+const checkoutController = require('../../controllers/checkout-controllers/checkoutController');
 const authenticate = require('../../middleware/authMiddleware');
