+++ src/routes/product-routes/productRoutes.js
@@ -1,7 +1,7 @@
 
 const express = require('express');
 const router = express.Router();
-const productController = require('../../controllers/product controllers/productController');
+const productController = require('../../controllers/product-controllers/productController');
 const authenticate = require('../../middleware/authMiddleware');
 const { authorize, isVendorOrAdmin } = require('../../middleware/roleMiddleware');
